requirements.txt
PyYAML>=6.0
psutil>=5.9.0
pytest>=7.0.0